server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
openai==0.28.1
httpx==0.24.1
pydantic==2.5.2
aiofiles==23.2.1
numpy==1.26.2
soundfile==0.12.1
librosa==0.10.1
redis==5.0.1
celery==5.3.6
aiohttp==3.9.1 